src/Ocelot/Configuration/Creator/HttpHandlerOptionsCreator.cs
namespace Ocelot.Configuration.Creator
{
    using Logging;
    using Microsoft.Extensions.DependencyInjection;
    using Ocelot.Configuration.File;
    using System;

    public class HttpHandlerOptionsCreator : IHttpHandlerOptionsCreator
    {
        private readonly ITracer _tracer;

        public HttpHandlerOptionsCreator(IServiceProvider services)
        {
            _tracer = services.GetService<ITracer>();
        }

        public HttpHandlerOptions Create(FileHttpHandlerOptions options)
        {
            var useTracing = _tracer != null && options.UseTracing;

            return new HttpHandlerOptions(options.AllowAutoRedirect,
                options.UseCookieContainer, useTracing, options.UseProxy);
        }
    }
}
